Added Binary and Hex convertion & edited _printf-f
