+++ _printf.c
@@ -3,8 +3,8 @@
 int _printf(char *format, ...)
 {
 	int times, i;
-	va_list args;
 
+	va_list args;
 	va_start(args, format);
 	for (i = 0; format[i] != '\0'; i++)
 	{
@@ -13,24 +13,17 @@ int _printf(char *format, ...)
 			i++;
 			switch (format[i])
 			{
-				/* Case for intiger and decimal */
 			case 'i':
 			case 'd':
 			{
-				int arg = va_arg(args, int);
-
-				times += print_int(arg);
+				times += print_int(args);
 				break;
 			}
-			/* Case on string */
 			case 's':
 			{
-				char *arg = va_arg(args, char *);
-
-				times += print_string(arg);
+				times += print_string(args);
 				break;
 			}
-			/* Case on Characters */
 			case 'c':
 			{
 				char arg = va_arg(args, int);
@@ -38,12 +31,19 @@ int _printf(char *format, ...)
 				times += print_char(arg);
 				break;
 			}
-			/* Case on Character % */
 			case '%':
 			{
 				times += print_char('%');
 				break;
 			}
+			case 'b':
+			{
+				times += print_char('%');
+				break;
+			}
+			default:
+				times += write(1, &format[i - 1], 2);
+				break;
 			}
 		}
 		else
@@ -57,14 +57,15 @@ int _printf(char *format, ...)
 }
 
 /* Completed functions */
-int print_char(char g)
+int print_char(char arg)
 {
-	return (write(1, &g, 1));
+	return (write(1, &arg, 1));
 }
 
 /* Completed function */
-int print_int(int arg)
+int print_int(va_list args)
 {
+	int arg = va_arg(args, int);
 	char buffer[12];
 	int i = 0;
 	int count = 0;
@@ -100,16 +101,16 @@ int print_int(int arg)
 }
 
 /* Completed function */
-int print_string(char *string)
+int print_string(va_list args)
 {
 	int len;
-
-	len = strlen(string);
-	if (string == NULL)
+	char *arg = va_arg(args, char *);
+	len = strlen(arg);
+	if (arg == NULL)
 	{
-		string = "(null)";
-		len = strlen(string);
+		arg = "(null)";
+		len = strlen(arg);
 	}
 
-	return (write(1, string, len));
+	return (write(1, arg, len));
 }
