+++ test/discarded/more_functions.h
